Made class ErrorHandler the parent for all classes, made the transition structure of input through all checks.
Made the ErrorHandler the parent class and the other classes the children classes, which inherit it's methods of error-checking, in order to make the program more structured and having all possible errors in one place. Also, made the structure of how the input will be checked for every level, syntactically and logically.