Improved README file, updated try-catch, updated imports.
In README it was added the structure, the team working on the project and also the input flow (which is in progress) with some brainstorming ideas. Updated imports, by removing the semi-column at the final for not having C4067 warning and also removing the ErrorHandler as a class parent by making the try-catch statement for errors.